requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.11.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.346
langchain-core==0.0.10
langsmith==0.0.69
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.7
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==4.1.0
protobuf==4.25.1
pyarrow==14.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
ruff==0.1.7
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
streamlit==1.29.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
yarl==1.9.4
zipp==3.17.0
